added special precipitation data
